+++ Cargo.toml
@@ -32,7 +32,6 @@ rattler_lock = "0.22.42"
 rattler_conda_types = { version = "0.30.2", default-features = false, features = [
   "rayon",
 ] }
-py-rattler = { git = "https://github.com/conda/rattler", tag = "rattler-v0.32.0" }
 
 dirs = "6.0.0"
 fs-err = "3.1.0"
